requirements.txt
azure-identity
azure-keyvault-secrets
tqdm
prettytable
azure-mgmt-compute
azure-mgmt-network
azure-mgmt-storage
azure-mgmt-resource
tabulate